LeadManagementsys/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-import-export==3.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
pycodestyle==2.10.0
PyMySQL==1.0.2
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
tzdata==2022.6
xlrd==2.0.1
xlwt==1.3.0